+++ keypair/exchange/exchange.go
@@ -52,7 +52,7 @@ func (p *PublicKey) String() string {
 	status := C.self_exchange_public_key_encode(
 		(*C.self_exchange_public_key)(p),
 		(*C.uint8_t)(unsafe.Pointer(&encoded[0])),
-		C.ulong(len(encoded)),
+		C.size_t(len(encoded)),
 	)
 
 	if status > 0 {
